requirements.txt
Flask==2.0.3
line-bot-sdk==1.20.0
gunicorn==20.1.0
